UlamMethod.jl

Documentation for UlamMethod.jl (test!)

UlamMethod.binner_square — Method

binner_square(domain)

Cover the computational domain in domain by a uniform grid of squares. Returns Vector{UlamPolygon{Float64}}.

UlamMethod.binner_voronoi — Method

binner_voronoi(traj, domain; rseed)

Cover the domain by a Voronoi tesselation based on points generated by kmeans clustering of the initial trajectory points. Returns Vector{UlamPolygon{Float64}}.

UlamMethod.inpoly — Method

inpoly(data::Matrix{Float64}, polys::PolyTable)

Determines which polygon of polys contains the data points in data. Returns an InpolyResult[@ref].

UlamMethod.inpoly — Method

inpoly(traj::UlamTrajectories, domain::UlamDomain)

Determines the indices of points that are inside the domain (i.e. not nirvana.)

UlamMethod.ulam_binner — Method

ulam_binner(traj, domain)

Selects and executes the appropriate binning algorithm for the given trajectories and domain. Intersects the resulting polygons with the domain so that the returned polygons are clipped to it. Returns Vector{UlamPolygon{Float64}}.

UlamMethod.ulam_intersection — Method

ulam_intersection(poly1, poly2)

Compute the intersection of two UlamPolygons objects. Returns false if they do not intersect.

UlamMethod.ulam_intersects — Method

ulam_intersects(poly1, poly2)

Return truw or false accoring to whether two UlamPolygons objects intersect. Can be faster than ulam_intersection if the shape of the intersection is not needed.

UlamMethod.ulam_method — Method

ulam_method(traj, domain)

The high-level Ulam method; cover the domain by polygons and then construct the transition probability matrix.

UlamMethod.ulam_nirvana — Method

ulam_nirvana(traj, domain, polys)

This is the implementation of Ulam's method with a nirvana state. Returns an UlamResult object. Returns an UlamResult