API

These are the full docstrings for UlamMethod.jl.

Index

UlamMethod.EarthPolygons.GoG_big

UlamMethod.EarthPolygons.GoG_small

UlamMethod.EarthPolygons.North_Atlantic_box

UlamMethod.EarthPolygons.North_Atlantic_clipped

UlamMethod.BinningAlgorithm

UlamMethod.Bins

UlamMethod.Boundary

UlamMethod.DataReinjection

UlamMethod.HexagonBinner

UlamMethod.HyperRectangle

UlamMethod.HyperRectangleBinner

UlamMethod.LineBinner

UlamMethod.RectangleBinner

UlamMethod.ReinjectionAlgorithm

UlamMethod.SourceReinjection

UlamMethod.Trajectories

UlamMethod.TriangleBinner

UlamMethod.UlamResult

UlamMethod.VoronoiBinner

UlamMethod.P_closed

UlamMethod.P_open

UlamMethod.bins

UlamMethod.bins_dis

UlamMethod.membership

UlamMethod.points

UlamMethod.ulam_method

Functions

UlamMethod.P_closed — Method

P_closed(ulam_result)

Return the full transition matrix.

UlamMethod.P_open — Method

P_open(ulam_result)

Return ulam_result.O2O, i.e. the transition matrix without nirvana.

UlamMethod.bins — Method

bins(ulam_result)

Return the bins associated to the transition matrix.

UlamMethod.bins_dis — Method

bins_dis(ulam_result)

Return the bins that contained data but were removed when the largest strongly connected component was taken.

UlamMethod.membership — Method

membership(data, ulam_result)

Takes a Dim x N_points matrix of points and returns a vector memb where memb[i] = j if data[:,i] is inside ulam_result.binner.bins[j] and memb[i] = nothing if data[:,i] is not inside any bin.

UlamMethod.points — Method

points(bins)

Return a vector of raw vertices for each bin. Can also be applied to an UlamResult.

UlamMethod.ulam_method — Method

ulam_method(traj, binner; reinj_algo)

Run the main Ulam's method calculation and return an UlamResult.

Arguments

traj: A Trajectories object, holding the short-range trajectory data.

binner: A BinningAlgorithm that specifies the algorithm used to partition the boundary into bins.

Optional Arguments

reinj_algo: A ReinjectionAlgorithm that specifies how trajectories pointing from nirvana to the interior should be reinjected. Default DataReinjection.

Types

UlamMethod.BinningAlgorithm — Type

abstract type BinningAlgorithm{Dim}

An abstract type for binning algorithms of dimension Dim.

Implementation

Each subtype should have a fields:

boundary::Boundary{Dim, CRS}`

bins::Bins{Dim, CRS}

idx2pos::Vector{Union{Int64, Nothing}}

Each subtype should implement the following:

A function membership(data::Matrix, binner) which takes a Dim x N_points matrix of points and returns a vector memb where memb[i] = j if data[:,i] is inside binner.bins[j] and memb[i] = nothing if data[:,i] is not inside any bin.

A function membership(traj::Trajectories, binner) which returns (memb_x0, memb_xT). In many cases this will be given by (membership(traj.x0, binner), membership(traj.xT, binner)).

Methods

Use points(binner) to return a vector of raw vertices for each bin.

UlamMethod.Bins — Type

struct Bins{Dim, CRS}

A container type for bins.

Fields

bins: A vector of Polytope objects of dimension Dim and coordinate reference system CRS.

Methods

Use points(bins) to return a vector of raw vertices for each bin.

UlamMethod.Boundary — Type

struct Boundary{Dim, CRS}

A container type for the computational domain within which Ulam's method is applied.

The boundary is partitioned according to an BinningAlgorithm.

Fields

boundary: A Polytope defining the boundary.

1D

The boundary is a continuous line segment Segment between x_start and x_end. Use

Boundary(x_start, x_end)

Use points(boundary) to calculate a Dim x N matrix of boundary vertices.

2D

The boundary is a closed polygon Ngon with vertices verts, which should be be a vector of (x, y) coordinates or a 2 x N matrix. Use

Boundary(verts)

A convenience constructor is also provided for the case of a rectangular boundary. Use

Boundary(xmin, xmax, ymin, ymax)

Use points(boundary) to calculate a Dim x N matrix of boundary vertices.

≥3D

The boundary is a HyperRectangle with minimum and maximum vertices corner_min, corner_max. Use

Boundary(corner_min, corner_max)

where the corners are NTuples, (x_min, y_min, z_min, ...), (x_max, y_max, z_max, ...).

UlamMethod.DataReinjection — Type

struct DataReinjection

Reinject the data according to which bins are actually hit by transitions from the exterior to the interior.

Since this is applied by default in the main Ulam's method algorithm, the associated reinject! function for DataReinjection does nothing.

Constructor

DataReinjection()

UlamMethod.HexagonBinner — Type

struct HexagonBinner{CRS}

Bin a two dimensional Polygon with a covering of regular hexagons.

The final number of bins may be slightly different than the number requested.

Fields

boundary: A Boundary object.

bins: A Bins object.

idx2pos: A vector such that idx2pos[i] gives the position (in bins) of the bin initially (before removing dataless and disconnected bins) labelled i. idx2pos[i] == nothing if this bin was removed.

Constructor

HexagonBinner(nbins, boundary; hardclip = true)

UlamMethod.HyperRectangle — Type

struct HyperRectangle{K, Dim, CRS}

An extention of Meshes to dimensions greater than 3.

The extension is minimal and implements no methods and does not interface with CRS.

Fields

min: The coordinates of the minimum of the HyperRectangle as an NTuple.

max: The coordinates of the maximum of the HyperRectangle as an NTuple.

vertices: [min, max], required by Meshes.

Constructor

HyperRectangle(min, max)

UlamMethod.HyperRectangleBinner — Type

struct HyperRectangleBinner{Dim, CRS}

Bin Dim-dimensional (where Dim ≥ 3) dataset based on a tight covering of identical HyperRectangles.

The hyperrectangles are chosen to be as close as possible to hypercubes.

The HyperRectangleBinner does not have hard clipping functionality. However, bins with no data are still removed in the main Ulam's method calculation.

The final number of bins may be different (larger) than the number requested.

Fields

boundary: A Boundary object.

bins: A Bins object.

idx2pos: A vector such that idx2pos[i] gives the position (in bins) of the bin initially (before removing dataless and disconnected bins) labelled i. idx2pos[i] == nothing if this bin was removed.

ranges: A vector such that ranges[i] is the AbstractRange giving the gridpoints of the bins along dimension i.

Constructor

HyperRectangleBinner(nbins, boundary)

UlamMethod.LineBinner — Type

struct LineBinner{CRS}

Bin a one dimensional Segment (line segement) with nbins equally-spaced bins.

Fields

boundary: A Boundary object.

bins: A Bins object.

idx2pos: A vector such that idx2pos[i] gives the position (in bins) of the bin initially (before removing dataless and disconnected bins) labelled i. idx2pos[i] == nothing if this bin was removed.

Constructor

LineBinner(nbins, boundary; hardclip = true)

UlamMethod.RectangleBinner — Type

struct RectangleBinner{CRS}

Bin a two dimensional Polygon with a tight covering of rectangles. The rectangles are chosen to be as close as possible to squares and so the final number of bins may be slightly different than the number requested.

Fields

boundary: A Boundary object.

bins: A Bins object.

idx2pos: A vector such that idx2pos[i] gives the position (in bins) of the bin initially (before removing dataless and disconnected bins) labelled i. idx2pos[i] == nothing if this bin was removed.

Constructor

RectangleBinner(nbins, boundary; hardclip = true)

UlamMethod.ReinjectionAlgorithm — Type

abstract type ReinjectionAlgorithm

The abstract type for the algorithm controlling how trajectories pointing from nirvana to the interior are redistributed.

Each subtype reinj_algo should implement the following:

a function reinject!(binner, Pij, reinj_algo) that modifies the transition matrix Pijin place given

the binning algorithm binner. This function is applied after the largest strongly connected component of the original Pij is taken, but before the matrix is row-stochasticized.

UlamMethod.SourceReinjection — Type

struct SourceReinjection{Dim}

Reinject the data at particular locations.

Fields

points: A vector of the form [p1, p2, ...] where pi is a point such as (x) or (x, y). Data are reinjected uniformly across all bins that contain at least one member of points.

fallback: The ReinjectionAlgorithm to apply in case no bins contain any points or there is no data to reinject.

Constructor

SourceReinjection(points; fallback = DataReinjection())

UlamMethod.Trajectories — Type

struct Trajectories{Dim}

A container for trajectory data of dimension Dim.

Fields

x0: Initial coordinates, should be a matrix of size Dim x n_traj.

xT: Final coordinates, should be a matrix of size Dim x n_traj.

Constructor

Trajectories(x0, xT)

UlamMethod.TriangleBinner — Type

struct TriangleBinner{CRS}

Bin a two dimensional Polygon with a covering of equilateral triangles.

The final number of bins may be slightly different than the number requested.

Fields

boundary: A Boundary object.

bins: A Bins object.

idx2pos: A vector such that idx2pos[i] gives the position (in bins) of the bin initially (before removing dataless and disconnected bins) labelled i. idx2pos[i] == nothing if this bin was removed.

Constructor

TriangleBinner(nbins, boundary; hardclip = true)

UlamMethod.UlamResult — Type

struct UlamResult{Dim, CRS}

A container for the result of the main Ulam's method calculation.

This consists of three main components, the transition probability matrix, the associated bins and the bins that were disconnected.

The transition probability matrix is of the form

| PO2O | PO2ω |

| P_ω2O | 0 |

Fields

P_O2O: As in diagram. P_O2ω: As in diagram. P_ω2O: As in diagram. binner: The [BinningAlgorithm] used to bin the data. bins_dis: The bins that contained data but were removed when the largest strongly connected component was taken.

Methods

Use P_open(UlamResult) to access P_O2O and P_closed(UlamResult) to access the full matrix.

Use bins(UlamResult) and bins_dis(UlamResult) to access bins and bins_dis, respectively.

UlamMethod.VoronoiBinner — Type

struct VoronoiBinner{Dim, CRS}

Bin Dim-dimensional (where Dim ∈ (1, 2)) dataset based on a Voronoi tesellation of points generated by a k-means clustering of initial trajectory data (i.e. Trajectories.x0).

Fields

boundary: A Boundary object.

bins: A Bins object.

idx2pos: A vector such that idx2pos[i] gives the position (in bins) of the bin initially (before removing dataless and disconnected bins) labelled i. idx2pos[i] == nothing if this bin was removed.

Constructor

VoronoiBinner(nbins, boundary, traj; hardclip = true)

<!– ## Constants

–>

Earth Polygons

UlamMethod.EarthPolygons.GoG_big — Constant

GoG_big

Vertices defining the entirety of the Gulf of Guinea.

UlamMethod.EarthPolygons.GoG_small — Constant

GoG_small

Vertices defining the interior of the Gulf of Guinea closest to the shore.

UlamMethod.EarthPolygons.North_Atlantic_box — Constant

North_Atlantic_box

Vertices defining a rectangle with lower left corner (-100, -9) and upper right corner (15, 39).

UlamMethod.EarthPolygons.North_Atlantic_clipped — Constant

North_Atlantic_clipped

Vertices defining a polygon contained in the rectangle with lower left corner (-100, -9) and upper right corner (15, 39). This polygon clips away the land, leaving only the ocean in the rectangle.