Full Reference

This is a full reference for UlamMethod.jl.

Index

UlamMethod.binner_square

UlamMethod.binner_voronoi

UlamMethod.inpoly

UlamMethod.inpoly

UlamMethod.ulam_binner

UlamMethod.ulam_intersection

UlamMethod.ulam_intersects

UlamMethod.ulam_nirvana

Items

UlamMethod.binner_square — Method

binner_square(domain)

Cover domain by a tight uniform grid of squares.

UlamMethod.binner_voronoi — Method

binner_voronoi(traj, domain)

Cover the domain by a Voronoi tesselation based on points generated by kmeans clustering of the initial trajectory points.

UlamMethod.inpoly — Method

inpoly(data::Matrix{Float64}, polys::PolyTable)

Determine which polygon of polys contains the data points in data. Return an InpolyResult.

UlamMethod.inpoly — Method

inpoly(traj::UlamTrajectories, domain::UlamDomain)

Determine the indices of points that are inside the domain's rectangle, but outside domain.domain. Return an InpolyResult[@ref].

UlamMethod.ulam_binner — Method

ulam_binner(traj, domain)

Select and executes the appropriate binning algorithm for the given trajectories and domain.

Intersect the resulting polygons with the domain so that the returned polygons are clipped to it.

UlamMethod.ulam_intersection — Method

ulam_intersection(poly1, poly2)

Compute the intersection of two UlamPolygon objects. Return false` if they do not intersect.

UlamMethod.ulam_intersects — Method

ulam_intersects(poly1, poly2)

Return true or false accoring to whether two [UlamPolygon](@ref) objects intersect. Can be faster than [ulam_intersection`](@ref) if the shape of the intersection is not needed.

UlamMethod.ulam_nirvana — Method

ulam_nirvana(traj, domain, polys)

Execute the core algorithm of Ulam's method with a nirvana state. Return an UlamResult.