API

These are the methods and types used by UlamMethod.jl.

Index

UlamMethod.global_poly_number_default

UlamMethod.global_poly_types

UlamMethod.global_rseed_default

UlamMethod.global_stoc_types

UlamMethod.global_traj_file_types

UlamMethod.InpolyResult

UlamMethod.PolyTable

UlamMethod.PolyTable

UlamMethod.UlamDomain

UlamMethod.UlamDomain

UlamMethod.UlamDomain

UlamMethod.UlamInfo

UlamMethod.UlamPolygon

UlamMethod.UlamPolygon

UlamMethod.UlamResult

UlamMethod.UlamResult

UlamMethod.UlamTrajectories

UlamMethod.UlamTrajectories

UlamMethod.UlamTrajectories

UlamMethod.binner_rectangle

UlamMethod.binner_square

UlamMethod.binner_voronoi

UlamMethod.inpoly

UlamMethod.inpoly

UlamMethod.ulam_binner

UlamMethod.ulam_intersection

UlamMethod.ulam_intersects

UlamMethod.ulam_method

UlamMethod.ulam_nirvana

UlamMethod.ulam_write

Functions

UlamMethod.binner_rectangle — Method

binner_rectangle(domain)

Cover the rectangle defined by domain.corners by a tight uniform grid of rectangles.

The rectangles are chosen to be as close to squares as possible with the constraint that they must exactly fit domain.corners.

UlamMethod.binner_square — Method

binner_square(domain)

Cover the rectangle defined by `domain.corners`by a uniform grid of squares.

The squares near the edges may overlap with domain.corners depending on the number requested.

UlamMethod.binner_voronoi — Method

binner_voronoi(traj, domain)

Cover the rectangle defined by domain.corners by a Voronoi tesselation based on points generated by kmeans clustering of the initial trajectory points in traj.

UlamMethod.inpoly — Method

inpoly(data::Matrix{Float64}, polys::PolyTable)

Determine which polygon of polys contains the data points in data. Return an InpolyResult.

UlamMethod.inpoly — Method

inpoly(traj::UlamTrajectories, domain::UlamDomain)

Determine the indices of points that are inside domain.domain. Return an InpolyResult.

UlamMethod.ulam_binner — Method

ulam_binner(traj, domain)

Select and executes the appropriate binning algorithm for the given trajectories and domain.

Intersect the resulting polygons with the domain so that the returned polygons are clipped to it.

UlamMethod.ulam_intersection — Method

ulam_intersection(poly1, poly2)

Compute the intersection of two UlamPolygon objects. Return false if they do not intersect or if they only intersect along a point or line.

UlamMethod.ulam_intersects — Method

ulam_intersects(poly1, poly2)

Return true or false accoring to whether two UlamPolygon objects intersect.

This is a call to ulam_intersection, hence two polygons are not considered to intersect if their intersection is only along a point or line.

UlamMethod.ulam_method — Method

ulam_method(traj, domain)

Run the high-level Ulam method and return an UlamResult.

Arguments

traj: An UlamTrajectories; contains the trajectory data.

domain: An UlamDomain; contains the domain specification.

UlamMethod.ulam_nirvana — Method

ulam_nirvana(traj, domain, polys)

Execute the core algorithm of Ulam's method with a nirvana state. Return an UlamResult.

UlamMethod.ulam_write — Method

ulam_write(outfile, ulam_result; dir_name, overwrite, P_out)

Write ulam_result to the file outfile, which must be in the .h5 format.

If outfile does not exist, it will be created. Results are written to the directory specified by dir_name.

Optional Arguments

dir_name: The name of the directory that ulam_result is written to, default "ulam". Directories can be nested, e.g. dir_name = "trial1/ulam".

overwrite: If true, the directory dir_name will overwrite a directory with the same name if it exists in outfile. Default false.

P_out: If false, P_closed is not written to file. Default true.

Types

UlamMethod.InpolyResult — Type

InpolyResult{U}

Hold the result of inpoly.

UlamMethod.PolyTable — Type

Polytable{T, U}

Tables, in matrix form of nodes and edges making up a set of polygons.

UlamMethod.PolyTable — Method

Polytable(ulam_polys)

Construct a table of matrices and nodes of a vector of UlamPolygons.

UlamMethod.UlamDomain — Type

UlamDomain{S, T, U}

The domain defining the Ulam problem setup.

Fields

corners: The rectangular bounding box of the polygon in domain. Binning algorithms start by covering this rectangle.

domain: Data outside this are considered in nirvana.

poly_type: The kind of polygons which cover domain.

poly_number: The number of polygons which cover domain.

stoc_type: The manner in which trajectories which leave domain are reinjected.

stoc_source: The location in which to reinject data for the "source" reinjection algorithm.

UlamMethod.UlamDomain — Method

UlamDomain(domain; [...])

Construct an UlamDomain defined by the UlamPolygon in domain.

Points outside domain will be considered in nirvana.

Optional Arguments

poly_type: One of "rec", "sqr", "hex", and "vor" for coverings by rectangles, squares, hexagons or Voronoi tesselation. The default is rectangles.

poly_number: The number of polygons requested. The default is 500 for rectangles/squares/hexagons and 100 for Voronoi.

stoc_type: Picks the stochasticization algorithm; one of "data" or "source". The default is data.

stoc_source::UlamPolygon: Polygons in the covering that intersect stoc_source will have data re-injected uniformly through them in the source algorithm.

stoc_source::Matrix: Polygons in the covering which contain the points in the stoc_source matrix will have data re-injected uniformly through them in the source algorithm.

rseed: A seed for reproducing the random initialization of the kmeans algorithm in the Voronoi covering.

UlamMethod.UlamDomain — Method

UlamDomain(xmin, xmax, ymin, ymax; [...])

Construct an UlamDomain defined by the rectangle with bottom left corner (xmin, ymin) and top right corner (xmax, ymax).

This is equivalent to UlamDomain(domain; [...]) where the provided domain is a rectangular UlamPolygon.

UlamMethod.UlamInfo — Type

UlamInfo{S, U}

A container for some of the results of ulam_method.

UlamMethod.UlamPolygon — Type

UlamPolygon{T}

A polygon defined by a series of nodes and carrying a center.

UlamMethod.UlamPolygon — Method

UlamPolygon(nodes; edges = nothing)

Construct an UlamPolygon based on a nodes matrix with n rows and 2 columns.

The nodes should not have a closing node, i.e. the last node should not be a repeat of the first.

The edges of the polygon defined by nodes are connected in order.

Optional Arguments

edges: an n by 2 matrix which specifies the edge connections betwen nodes. Used if nodes are not already sorted.

UlamMethod.UlamResult — Type

UlamResult{S, T, U}

A container for the output of ulam_method.

UlamMethod.UlamResult — Method

UlamResult(P_closed, polys, polys_dis, counts, info)

Construct a container for the output of ulam_method.

Arguments

P_closed: the transition probability matrix obtained by Ulam's method.

polys: the vector of UlamPolygons constituting the covering of the domain.

polys_dis: the vector of UlamPolygons which contained data but were in disconnected components.

counts: the number of observation data points in the ith polygon.

info: additional info from UlamInfo.

UlamMethod.UlamTrajectories — Type

UlamTrajectories{T}

A container for trajectory data.

UlamMethod.UlamTrajectories — Method

UlamTrajectories(infile, [...])

Construct a container for trajectory data, loading it from infile, which should be a .h5 or .mat file.

Optional Arguments

These are used if trajectory data in infile are named something other than "x0", "y0", "xT", "yT".

x0_alias

y0_alias

xT_alias

yT_alias

UlamMethod.UlamTrajectories — Method

UlamTrajectories(;x0, y0, xT, yT)

Construct a container for trajectory data.

Constants

UlamMethod.global_poly_number_default — Constant

global_poly_number_default

The default number of polygons for each of the types of coverings.

"rec": 500

"sqr": 500

"hex": 500

"vor": 100

UlamMethod.global_poly_types — Constant

global_poly_types

One of "rec", "sqr", "hex" and "vor" according to the coverings in UlamDomain.

UlamMethod.global_rseed_default — Constant

global_rseed_default

The default seed for RNG reproducibility. Default 123.

UlamMethod.global_stoc_types — Constant

global_stoc_types

One of "data" or "source" according to the stochasticization options in UlamDomain.

UlamMethod.global_traj_file_types — Constant

global_traj_file_types

One of "mat", "h5" according to the trajectory input file in UlamTrajectories