Reference

UlamMethod.jl includes the following methods and types.

Index

UlamMethod.UlamTypes.InpolyResult

UlamMethod.UlamTypes.PolyTable

UlamMethod.UlamTypes.PolyTable

UlamMethod.UlamTypes.UlamDomain

UlamMethod.UlamTypes.UlamDomain

UlamMethod.UlamTypes.UlamInfo

UlamMethod.UlamTypes.UlamPolygon

UlamMethod.UlamTypes.UlamPolygon

UlamMethod.UlamTypes.UlamResult

UlamMethod.UlamTypes.UlamResult

UlamMethod.UlamTypes.UlamTrajectories

UlamMethod.UlamTypes.UlamTrajectories

UlamMethod.UlamTypes.UlamTrajectories

UlamMethod.ulam_method

Functions

UlamMethod.ulam_method — Method

ulam_method(traj, domain)

Run the high-level Ulam method and return an UlamResult.

Arguments

traj: An UlamTrajectories; contains the trajectory data.

domain: An UlamDomain; contains the domain specification.

Types

UlamMethod.UlamTypes.InpolyResult — Type

InpolyResult{U}

Hold the result of inpoly`.

UlamMethod.UlamTypes.PolyTable — Type

Polytable{T, U}

Tables, in matrix form of nodes and edges making up a set of polygons.

UlamMethod.UlamTypes.PolyTable — Method

Polytable(ulam_polys)

Construct a table of matrices and nodes of a vector of UlamPolygons.

UlamMethod.UlamTypes.UlamDomain — Type

UlamDomain{S, T, U}

The domain defining the Ulam problem setup.

UlamMethod.UlamTypes.UlamDomain — Method

UlamDomain(xmin, xmax, ymin, ymax; [...])

Construct an UlamDomain defined by the rectangle with bottom left corner (xmin, ymin) and top right corner (xmax, ymax).

Optional Arguments

domain: Points inside the rectangle, but outside the outside domain will be considered in nirvana. This can be used to refine the shape of the computational domain to an arbitrary UlamPolygon, not just the default rectangle.

poly_type: One of "sqr", "hex" and "vor" for coverings by squares, hexagons or Voronoi tesselation. The default is squares.

poly_number: The number of polygons requested. The default is 500 for squares/hexagons and 100 for Voronoi.

stoc_type: Picks the stochasticization algorithm; one of "data" or "source". The default is data.

stoc_polygon: Polygons in the covering that intersect stoc_polygon will have data re-injected uniformly through them in the source algorithm.

rseed: A seed for reproducing the random initialization of the kmeans algorithm in the Voronoi covering.

UlamMethod.UlamTypes.UlamInfo — Type

UlamInfo{S, U}

A contained for some of the results of ulam_method.

UlamMethod.UlamTypes.UlamPolygon — Type

UlamPolygon{T}

A polygon defined by a series of nodes and carrying a center.

UlamMethod.UlamTypes.UlamPolygon — Method

UlamPolygon(nodes; edges = nothing)

Construct an UlamPolygon based on a nodes matrix with n rows and 2 columns.

The edges of the polygon defined by nodes are connected in order.

Optional Arguments

edges: an n by 2 matrix which specifies the edge connections betwen nodes. Used if nodes are not already sorted.

UlamMethod.UlamTypes.UlamResult — Type

UlamResult{S, T, U}

A container for the output of ulam_method.

UlamMethod.UlamTypes.UlamResult — Method

UlamResult(P_closed, polys, polys_dis, counts, info)

Construct a container for the output of ulam_method.

Arguments

P_closed: the transition probability matrix obtained by Ulam's method.

polys: the vector of UlamPolygons constituting the covering of the domain.

polys_dis: the vector of UlamPolygons which contained data but were in disconnected components.

counts: the number of observation data points in the ith polygon.

info: additional info from UlamInfo.

UlamMethod.UlamTypes.UlamTrajectories — Type

UlamTrajectories{T}

A container for trajectory data.

UlamMethod.UlamTypes.UlamTrajectories — Method

UlamTrajectories(infile, [...])

Construct a container for trajectory data, loading it from infile, which should be a .h5 or .mat file.

### Optional Arguments
These are used if trajectory data in `infile` are named something other than `"x0"`, `"y0"`, `"xT"`, `"yT"`.
- `x0_alias`
- `y0_alias`
- `xT_alias`
- `yT_alias`

UlamMethod.UlamTypes.UlamTrajectories — Method

UlamTrajectories(;x0, y0, xT, yT)

Construct a container for trajectory data.