Reference

UlamMethod.jl includes the following methods and types.

Index

UlamMethod.UlamTypes.UlamPolygon

UlamMethod.UlamTypes.UlamPolygon

UlamMethod.ulam_method

Functions

UlamMethod.ulam_method — Method

ulam_method(traj, domain)

The high-level Ulam method; cover the domain by polygons and then construct the transition probability matrix.

Types

UlamMethod.UlamTypes.UlamPolygon — Type

UlamPolygon{T}

A polygon defined by a series of nodes and carrying a center.

UlamMethod.UlamTypes.UlamPolygon — Method

UlamPolygon(nodes; edges = nothing)

Construct an UlamPolygon based on a nodes matrix with n rows and 2 columns.

The edges of the polygon defined by nodes are connected in order.

Optional Arguments

edges::Matrix{<:Integer}: an n by 2 matrix which specifies the edge connections betwen nodes. Used if nodes are not already sorted.